task: reorganized substrait code
This reorganization is partially for my own benefit (readability and
internalizing the structure), though I did add some improvements such as
how substrait types are used in switch statements.

The other motivation for this reorganization is so that I can have a
better overall structure/framework for also "translating" a substrait
plan to a logical duckdb plan (instead of only a physical duckdb plan).

In this code, I use "translation" to refer to a _lowering_ from a
logical substrait plan to a physical duckdb plan.