task: added code for transpilation of substrait
In this code, which is an initial placeholder, I use "transpilation" to
refer to a translation (general sense of the word) from a logical
substrait plan to a logical duckdb plan. Thus, going forward I will be
using "translation" to refer to a lowering (e.g. logical to physical)
and "transpilation" to refer to a transformation to another language
(dialect?) of the same level (e.g. logical to logical).