task: mostly refactored style
Some re-ordering was done to group hash functors above the classes.
Also, `class` was changed to `struct` mostly for simplicity (not sure
why `SubstraitFunctionExtensions` was a `class` if everything was
public) and then a tiny amount for consistency (might as well maake
`SubstraitCustomFunctions` consistent with other classes).

Additionally, I added type aliases for templated unordered_map types for
readability