fix: working on a fix for container-overflow
I see a warning about container-overflow from lldb. It's possible it's a
false positive, but I want to try a different way of initializing the
`SubstraitCustomFunctions` registry to see if it helps. Unfortunately, I
started this in a separate repo instead of the submodule where I can
test the change directly.

Updates to scripts/generate_custom_functions.py was primarily for
readability and to simplify the code.

Changess to custom_extensions header and implementations is to change
the constructor away from populating the unordered_map attributes before
the object itself is constructed.