task: significant changes to CMakeLists.txt
The biggest changes stem from changing dependency on substrait from
in-tree source files to a separately built and installed library. The
point of this is to address duplicate descriptor issues as well as to
remove the vendored protobuf code.

Note that this change depends on a change in duckdb/CMakeLists.txt where
`build_loadable_extension` is defined.