task: add initial impl for SkyRel
This initial implementation hardcodes an small number of slices to be
used for a "full scan" of a SkyRel, but should appropriately accept an
explicit list of slice indices to scan.

This implementation represents receiving a substrait plan for execution
with duckdb and "grounding" (specifying the data access path) at
translation time. The grounding is done by explicitly translating a
SkyRel message to a `scan_arrows_file` TableFunction (defined in arrow
extension). By doing this translation eagerly, the generated duckdb plan
will directly access the filesystem for SkyRel data and bypass the
catalog. This means that the duckdb extension must have intimate
knowledge of the extension to be used. Essentially, the question is
whether intimate knowledge of the extension is implemented at substrait
translation time or at catalog lookup time.