fix: my fork does not have duckdb as a submodule
for now this means I can't add test/c as a cmake subdirectory. In the
long run, I probably can add it and duckdb-skytether repo should be
okay.

I should note that after rebasing a condition was added that checks for
an environment variable, but for now I don't feel like changing the
build to accommodate the variable. I should do this in the future though