task: updates to change substrait execution path
Instead of directly executing the duckdb::Relation constructed from the
substrait plan, it is translated to LogicalOperator then
PhysicalOperator then executed with a `duckdb::Executor`. The purpose of
this is to be able to always return a pushback plan instead of just a
ticket. This is necessary for cooperative decomposition.